Move some helper method from SerializationManager to another static class
